requirements.txt
flask
sentence_transformers
faiss-cpu
numpy
pandas